feat: add json schema generation
